ui: templates: Add validation of project name
Rewrite the home page project creator area. This is done in jquery as it
is a single button specific to the home page.

Make sure that we can only create compatible project names with the
project slugs. A global js object is created. Added to it is a regex
for testing project names and the editMode variable between pages.

Also remove a debug print from views.