ui: project: Re-work the auto project update mechanism
If a user enters changes that have yet to be saved such as
the new JSON key the auto-update project settings overwrites
the in-progress changes.

Add a flag unsavedChanges to pause updating the project properties from
the server until the changes are saved.

Add a ui component to remind users that there are unsaved changes.

Add mechanism to make sure the unsavedChanges flag is unset when saved and
unset when user reverts the changes.